refactor(data-store): extract sqlite-row-mappers from sqlite-adapter
Moves the row -> domain object mapping helpers out of sqlite-adapter into sqlite-row-mappers, behavior unchanged.
One atomic slice of the step-4 large-file decomposition.
